+++ contracts/src/v0.8/dev/vrf/testhelpers/VRFV2PlusSingleConsumerExample.sol
@@ -12,7 +12,7 @@ contract VRFV2PlusSingleConsumerExample is VRFConsumerBaseV2Plus {
   LinkTokenInterface LINKTOKEN;
 
   struct RequestConfig {
-    uint64 subId;
+    uint256 subId;
     uint32 callbackGasLimit;
     uint16 requestConfirmations;
     uint32 numWords;
